Refactor slice termination request
